requirements.txt
numpy==1.26.2
nibabel==5.1.0
monai==1.3.0
tqdm==4.66.1
tensorboard==2.15.1
lpips==0.1.4
scipy==1.11.4
torchmetrics==1.2.1
scikit-image==0.22.0
seaborn==0.13.0
scikit-learn==1.3.2
gdown==4.7.3
torcheval==0.0.7

git+https://github.com/Project-MONAI/GenerativeModels.git@894f2ec
